ExpParam

ExpParam holds all experimental data as well as database-related info generated during the processing steps. Use setExpParams to save it into your RHermesExp object

# S4 method for ExpParam
show(object)

Arguments

object
An ExpParam object

Value

A ExpParam object

Slots

ppm

The mass spectrometer (MS) mass error in parts per million (ppm)

res

The MS resolution at m/z = 200

nthr

The noise threshold to consider. Any signal weaker than it will NOT be considered in the PL. Useful to filter 'grass-like' peaks in a Q-TOF instrument. Defaults to 1000.

minmz

Minimum mz registered in the MS1 experiment. Defaults to 80

maxmz

Maximum mz registered in the MS1 experiment. Defaults to 1000.

ion

Polarity used in the experiment. RHermes currently only supports one polarity at a time per RHermesExp, so if you're performing both + and - polarity you will have to use 2 different RHermesExp objects, one for each. Important: must be described ONLY as '+' or '-'.

instr

MS instrument type used to acquire the data. It can be either QTOF' or 'Orbitrap'.

DB

Formula database

adlist

Adduct list

ionF

Ionic formula list. Formed by all formula-adduct combinations

isoList

Isotopic pattern exploration results.

See also

Examples

 #Default object
 exp <- ExpParam()
 #Setting some values - check all possible slots above
 exp2 <- ExpParam(ppm = 3, res = 150000, instr = "Orbitrap")