FileProc

Main function to process the input MS1 mzML files. It accesses the files with mzR and processes them with RHermes internal functions. For each file a PL object is generated inside the object_data_PL slot.

FileProc(struct, files, labelled = FALSE)

Arguments

struct
RHermesExp S4 object to update with the processed data. IMPORTANT: The objects needs to have the experimental parameters already set before calling this function.

files
Character vector of all the paths to the files to be processed.

labelled
Logical, whether to check for all 13C isotopic signals. Still in development. Defaults to FALSE

Value

An RHermesExp object with the processed PLs.

Examples

FileProc(myHermes, 'C:/myFolder/myFile.mzML', FALSE) #For regular use

#> Error in FileProc(myHermes, "C:/myFolder/myFile.mzML", FALSE): object 'myHermes' not found

FileProc(myHermes, 'C:/myFolder/myFile.mzML', TRUE) #For labelled data

#> Error in FileProc(myHermes, "C:/myFolder/myFile.mzML", TRUE): object 'myHermes' not found