ILParam

The ILParam class contains all details regarding IL generation, such as m/z window for grouping close SOIs, whether to prioritize or restrict the inclusion list to a subset of adducts, etc.

Value

An ILParam object

Details

It's a part of the RHermesIL object and is an input for generateIL().

Slots

filtermz

Deltamz used to group SOIs with similar RTs. Defaults to +-0.5Da. It should be matched with the m/z isolation window of your MS2 method settings.

filterrt

Delta RT used to group SOIs with similar RTs. Defaults to 10s.

rtmargin

Similar to filterrt. Defaults to 5s.

priorization

Whether to prioritze adducts or not. Options are "only" to select only a set of adduct annotations, "full" to cover all SOIs independently of the adduct annotations or "yes" to specify some adducts to be prioritized in order.

ad

The specified adducts. See priorization.

See also

Author

Roger Gine

Examples

ILParam()

#> An object of class "ILParam"
#> Slot "filtermz":
#> [1] 0.5
#> 
#> Slot "filterrt":
#> [1] 10
#> 
#> Slot "rtmargin":
#> [1] 5
#> 
#> Slot "priorization":
#> [1] "only"
#> 
#> Slot "ad":
#> [1] "M+H"
#>