MS2Proc

MS2Proc processes the MSMS files generated for a given IL and performs compound identification using an MSMS reference spectra database. Only the struct id and MS2files parameters are mandatory.

MS2Proc(
  struct,
  id,
  MS2files,
  referenceDB = "D:/sp_MassBankEU_20200316_203615.RData",
  mincos = 0.6,
  plotting = FALSE,
  outpdf = "mirror.pdf",
  outcsv = "./results.csv"
)

Arguments

struct
The RHermesExp object to add the processed data.

id
Numeric, the id of the inclusion list to process.

MS2files
Character vector of the MS2 files address.

referenceDB
Character, where to find the MSMS database to query against.

mincos
Numeric, between 0 and 1, the minimum spectral cosine between query and reference spectrum for a match to be reported (see the paper for more detail on this).

plotting
Logical, whether you want mirror plots to be generated.

outpdf
Character, where you want to store the plots (see plotting).

outcsv
Character, where to store a summary csv of the identifications.

Value

An RHermesExp object with the identifications set in the used MS2Exp slot.

Examples

MS2Proc(myHermes, 1, c('C:/myFolder/File1.mzML', 'C:/myFolder/File2.mzML'))

#> Error in MS2Proc(myHermes, 1, c("C:/myFolder/File1.mzML", "C:/myFolder/File2.mzML")): object 'myHermes' not found