SOIParam

The SOIParam class contains all details regarding SOI detection, such as the number of RT bins and their width, minimum data point intensity, usage of blank substraction, etc.

# S4 method for SOIParam
show(object)

Arguments

object
A SOIParam object

Value

A SOIParam object

Details

It's a part of the RHermesSOI object, and is an input for findSOI(). This object is usually generated by getSOIParam(), which features already-made templates.

Slots

specs

A data frame containing the RT bin information. Each row represents a binning step that will be applied to the annotated data points to detect the SOIs.

maxlen

Maximum SOI length in seconds. If the SOI is longer tha maxlen, a broad peak picking will divide the long SOI into smaller sized SOIs.

minint

Minimum data point intensity. If a point is smaller than minint, it will not be considered in the SOI detection.

blanksub

Logical. Whether to perform blanksub. It should NOT be entered by the user in this point, since findSOI will fill in this information.

blankname

Character. The corresponding blank file name. As before, it should not be entered by the user, findSOI will fill it automatically.

See also

Author

Roger Gine

Examples

SOIParam()

#> SOI parameters info:

#> 	Bins used:

#> data frame with 0 columns and 0 rows

#> 	Max SOI length:  30

#> 	Min data point intensity:  1000

#> 	Blank substraction performed:  FALSE

#> 	Blank filename:  None