addAd

Adds new custom adducts to the RHermesExp adduct list

addAd(struct, name, deltam, ch = 1, mult = 1, toadd = "FALSE", tosub = "FALSE")

# S4 method for RHermesExp,character,numeric
addAd(struct, name, deltam, ch = 1, mult = 1, toadd = "FALSE", tosub = "FALSE")

Arguments

struct
The RHermesExp object

name
Name of the adduct to add.

deltam
Delta m/z of the adduct, in Da.

ch
Adduct charge. Remember to specify the symbol if negative.

mult
Multiplicity (eg, M, 2M, 3M, etc.)

toadd
Atoms to add to the formula when it ionizes as the adduct. (ex: M+H -> H1, M+Cl -> Cl1). Note the explicit use of a 1 after the symbol.

tosub
Atoms to substract to the formula. Same idea as in toadd. (ex: M-H2O+H -> H2O1)

Value

An updated RHermesExp object with the new adduct list

Details

The function adds an entry to the Envipat-style dataframe containing the adducts used for the experiment. It is important to specify what atoms to add or substract to the molecular formula. Polarity is inferred directly from the specified charge.

Examples

if(FALSE){
 addAd(myHermes, 'M+H', 1.0072, ch = 1, mult = 1, toadd = 'H1')
}