generateIL

Groups the SOIs generated in the previous step and sets them up in an inclusion list ready for the MS/MS analysis. It takes into account the instrument precursor filter mz range and groups the SOIs accordingly. The user can also specify whether they want to use all SOIs or prioritize SOIs with a certain annotation (M+H adducts, prioritize SOIs with adequate isotopic patterns, etc.).

generateIL(struct, id, par)

# S4 method for RHermesExp,numeric
generateIL(struct, id, par = ILParam())

Arguments

struct
The RHermesExp object in which to save the resulting IL

id
The SOI list ID from which to generate the IL

par
An inclusion list parameter object, generated with ILParam()

Value

An RHermes object with the IL inside a new MS2Exp entry, inside the data slot

Examples

if(FALSE){
par <- ILParam(filtermz = 0.5, priorization = "full")
myHermes <- generateIL(myHermes, 1, par)
}