setTime

Adds a timestamp to the RHermesExp object with a specified message. Useful for adding info to the object for reproducibility purposes.

setTime(struct, message)

# S4 method for RHermesExp,character
setTime(struct, message)

Arguments

struct
The RHermesExp object you want to add the timestamp to.

message
A character string to append.

Value

A copy of struct with the updated timestamp list.

Author

Roger Gine

Examples

setTime(RHermesExp(), "Important info")

#> Experimental parameters info:

#> 	ppm error: 5

#> 	Mass resolution: 60000

#> 	Mass range: 80-1000

#> 	Noise cutoff value: 1000

#> 	Ionization mode: +

#> 	Number of compounds: 

#> 	Number of unique formulas: 0

#> 	Number of adducts: 0

#> 

#> No files processed yet.

#> Processing Timestamps ---------------------------------------

#> System info: Windows/10 x64/build 19041/LAPTOP-CFQ0D7GG/x86-64/Roger G/Roger G/Roger G

#> RHermes version: 0.99.0

#> Tue Mar 09 18:33:05 2021 - Important info

#>